refactor(protocol): change attribute metadata fields to f32 for float config support
Change decay_per_tick, default_value, min_value from i32 to f32 in
AttributeMetadata to support fractional attribute values (e.g. decay
rate of 0.5 per tick). Update all upstream/downstream code and remove
dead *_as_i32() accessor methods.

+++ crates/server/src/models/state_mutation.rs
@@ -64,7 +64,7 @@ impl AgentState {
 
             // 获取季节修饰系数（数据驱动）
             let season_modifier = self.get_season_modifier(&attr_name, tick_id);
-            let delta = (base_delta as f32 * season_modifier).floor() as i32;
+            let delta = (base_delta * season_modifier).floor() as i32;
 
             // 记录衰减前的值
             let before_value = self.status.get(&attr_name).unwrap_or(-1);
